requirements.txt
beautifulsoup4
requests
pandas
tabula-py
pyodbc
python-dotenv
lxml
